Write integration test for ldr_to_pfm and update luminosity limit
